Move hook into `selectors.ts`
