+++ web-admin/src/components/projects/selectors.ts
@@ -1,9 +1,43 @@
-import { createAdminServiceGetProject } from "@rilldata/web-admin/client";
+import {
+  V1DeploymentStatus,
+  createAdminServiceGetProject,
+} from "@rilldata/web-admin/client";
 
 export function getProjectPermissions(orgName: string, projName: string) {
   return createAdminServiceGetProject(orgName, projName, {
     query: {
       select: (data) => data?.projectPermissions,
+    },
+  });
+}
+
+const PollTimeDuringReconcile = 1000;
+const PollTimeDuringError = 5000;
+const PollTimeWhenProjectReady = 60 * 1000;
+
+export function useProjectDeploymentStatus(orgName: string, projName: string) {
+  return createAdminServiceGetProject<V1DeploymentStatus>(orgName, projName, {
+    query: {
+      select: (data) => {
+        return data?.prodDeployment?.status;
+      },
+      refetchInterval: (data) => {
+        switch (data) {
+          case "DEPLOYMENT_STATUS_PENDING":
+          case "DEPLOYMENT_STATUS_RECONCILING":
+            return PollTimeDuringReconcile;
+
+          case "DEPLOYMENT_STATUS_ERROR":
+          case "DEPLOYMENT_STATUS_UNSPECIFIED":
+            return PollTimeDuringError;
+
+          case "DEPLOYMENT_STATUS_OK":
+            return PollTimeWhenProjectReady;
+
+          default:
+            return PollTimeWhenProjectReady;
+        }
+      },
     },
   });
 }
